Restructured files, modularized schema
